+++ pkg/detector/detector_test.go
@@ -1,10 +1,12 @@
 package detector
 
 import (
+	"context"
 	"strconv"
 	"testing"
 	"time"
 
+	"github.com/giantswarm/micrologger"
 	"github.com/google/go-cmp/cmp"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -355,7 +357,9 @@ func Test_maximumNodeTermination(t *testing.T) {
 	}
 }
 
-func Test_nodeNotReady(t *testing.T) {
+func Test_isNodeUnhealthy(t *testing.T) {
+	const diskFullCondition corev1.NodeConditionType = "DiskFullKubelet"
+
 	testCases := []struct {
 		name                 string
 		node                 corev1.Node
@@ -392,7 +396,7 @@ func Test_nodeNotReady(t *testing.T) {
 			expectedNodeNotReady: true,
 		},
 		{
-			name: "test 1 - not ready but only for short time",
+			name: "test 2 - not ready but only for short time",
 			node: corev1.Node{
 				Status: corev1.NodeStatus{
 					Conditions: []corev1.NodeCondition{
@@ -406,13 +410,55 @@ func Test_nodeNotReady(t *testing.T) {
 			},
 			expectedNodeNotReady: false,
 		},
+		{
+			name: "test 3 - ready but disk full",
+			node: corev1.Node{
+				Status: corev1.NodeStatus{
+					Conditions: []corev1.NodeCondition{
+						{
+							Type:              corev1.NodeReady,
+							Status:            corev1.ConditionTrue,
+							LastHeartbeatTime: metav1.Time{Time: time.Now().Add(-time.Minute * 10)},
+						},
+						{
+							Type:              diskFullCondition,
+							Status:            corev1.ConditionTrue,
+							LastHeartbeatTime: metav1.Time{Time: time.Now().Add(-time.Minute * 10)},
+						},
+					},
+				},
+			},
+			expectedNodeNotReady: true,
+		},
+		{
+			name: "test 4 - ready but disk full for a short time",
+			node: corev1.Node{
+				Status: corev1.NodeStatus{
+					Conditions: []corev1.NodeCondition{
+						{
+							Type:              corev1.NodeReady,
+							Status:            corev1.ConditionTrue,
+							LastHeartbeatTime: metav1.Time{Time: time.Now().Add(-time.Minute * 10)},
+						},
+						{
+							Type:              diskFullCondition,
+							Status:            corev1.ConditionTrue,
+							LastHeartbeatTime: metav1.Time{Time: time.Now().Add(-time.Second * 10)},
+						},
+					},
+				},
+			},
+			expectedNodeNotReady: false,
+		},
 	}
 
 	for i, tc := range testCases {
 		t.Run(strconv.Itoa(i), func(t *testing.T) {
 			t.Log(tc.name)
 
-			result := nodeNotReady(tc.node)
+			logger, _ := micrologger.New(micrologger.Config{})
+
+			result := isNodeUnhealthy(context.Background(), logger, tc.node)
 			if result != tc.expectedNodeNotReady {
 				t.Fatalf("Expected '%t' but got '%t'.\n", tc.expectedNodeNotReady, result)
 			}
@@ -574,7 +620,9 @@ func Test_nodeNotReadyTickCount(t *testing.T) {
 		t.Run(strconv.Itoa(i), func(t *testing.T) {
 			t.Log(tc.name)
 
-			tickCounter, updated := nodeNotReadyTickCount(tc.node)
+			logger, _ := micrologger.New(micrologger.Config{})
+
+			tickCounter, updated := nodeNotReadyTickCount(context.Background(), logger, tc.node)
 			if tickCounter != tc.expectedTickCount {
 				t.Fatalf("Expected tick counter '%d' but got '%d'.\n", tc.expectedTickCount, tickCounter)
 			}