+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Consider `DiskFull` custom node conditions created by `node-problem-detector` app to check if a node is unhealthy.
+
 ## [2.0.0] - 2022-05-16
 
 ### Changed