+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.0.0] - 2023-11-09
+
 ### Added
 
 - Consider `DiskFull` custom node conditions created by `node-problem-detector` app to check if a node is unhealthy.
@@ -32,7 +34,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add github workflows for automatic releases.
 - Add bad node detector code.
 
-[Unreleased]: https://github.com/giantswarm/badnodedetector/compare/v2.0.0...HEAD
+[Unreleased]: https://github.com/giantswarm/badnodedetector/compare/v3.0.0...HEAD
+[3.0.0]: https://github.com/giantswarm/badnodedetector/compare/v2.0.0...v3.0.0
 [2.0.0]: https://github.com/giantswarm/badnodedetector/compare/v1.0.1...v2.0.0
 [1.0.1]: https://github.com/giantswarm/badnodedetector/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/giantswarm/badnodedetector/releases/tag/v1.0.0